requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# MCP SDK
mcp-sdk>=0.1.0

# HTTP Client
aiohttp>=3.8.0
requests>=2.26.0

# HTML Parser
beautifulsoup4>=4.9.3

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.17.0  # 异步SQLite支持
alembic>=1.7.0  # 数据库迁移工具

# Utils
pyyaml>=5.4.1
python-dotenv>=0.19.0
pydantic>=1.8.2

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.1

# Documentation
mkdocs>=1.2.3
mkdocs-material>=8.0.0
